separated utility functions into own dir
