+++ config/passport_utils/index.js
@@ -1,0 +1,8 @@
+const authCallback = require('./authCallback');
+const { serializeUser, deserializeUser } = require('./serializers');
+
+module.exports = {
+  authCallback,
+  serializeUser,
+  deserializeUser,
+};
